temporary save

+++ app/src/main/java/b1g4/com/yourseat/findPath/RecommendPath.java
@@ -86,7 +86,7 @@ public class RecommendPath {
         return false;
     }
 
-    public boolean isNearestStop( int startXPos, int startYPos ){
+    public boolean isNearestStop( double startXPos, double startYPos ){
 
         ArrayList<String> stationListBeforeStop = new ArrayList<String>();
         ArrayList<Double> distanceByDistance = new ArrayList<Double>();
